Propagate OpenRAVE core libraries to downstreams
The core component is a *superset* of base OR that loads the
OpenRAVE::RaveCreateEnvironment free function. We did not use it
in our launcher scripts on *runtime*, therefore it's not throwing
linker errors for now, but could fail under certain circumstances.
Also, propagate all necessary library dependencies through the
interface target, just to be sure.